+++ projects/packages/external-connections/src/class-external-connections.php
@@ -16,7 +16,7 @@ use Automattic\Jetpack\Status\Host;
  */
 class External_Connections {
 
-	const PACKAGE_VERSION = '0.1.9';
+	const PACKAGE_VERSION = '0.1.10';
 	const BASE_FILE       = __FILE__;
 
 	/**