+++ projects/packages/jetpack-mu-wpcom/src/features/help-center/class-wp-rest-help-center-persisted-open-state.php
@@ -63,9 +63,9 @@ class WP_REST_Help_Center_Persisted_Open_State extends \WP_REST_Controller {
 
 		$response = json_decode( wp_remote_retrieve_body( $body ) );
 
-		$is_open        = $response->help_center_open ?? false;
-		$is_minimized   = $response->help_center_minimized ?? false;
-		$router_history = $response->help_center_router_history ?? null;
+		$is_open        = $response->calypso_preferences->help_center_open ?? false;
+		$is_minimized   = $response->calypso_preferences->help_center_minimized ?? false;
+		$router_history = $response->calypso_preferences->help_center_router_history ?? null;
 
 		$projected_response = array(
 			'help_center_open'           => (bool) $is_open,