+++ projects/packages/explat/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.4] - 2025-12-22
+### Changed
+- Update dependencies. [#46381]
+
 ## [0.4.3] - 2025-12-11
 ### Changed
 - Update package dependencies. [#46244]
@@ -262,6 +266,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ExPlat: add condition to prevent fetching the experiment assignment if there's not anon id (meaning that Tracks is likely disabled) [#38327]
 - Updated package dependencies. [#38132]
 
+[0.4.4]: https://github.com/Automattic/jetpack-explat/compare/v0.4.3...v0.4.4
 [0.4.3]: https://github.com/Automattic/jetpack-explat/compare/v0.4.2...v0.4.3
 [0.4.2]: https://github.com/Automattic/jetpack-explat/compare/v0.4.1...v0.4.2
 [0.4.1]: https://github.com/Automattic/jetpack-explat/compare/v0.4.0...v0.4.1