+++ projects/packages/videopress/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.33.0] - 2025-12-22
+### Changed
+- Update package dependencies. [#46362] [#46363]
+- Make Gutenberg block depend on Jetpack site connection instead of user connection. [#46274]
+
 ## [0.32.19] - 2025-12-15
 ### Changed
 - Replace use of confusing `esc_js` with `wp_json_encode`. [#46229]
@@ -1801,6 +1806,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Created empty package [#24952]
 
+[0.33.0]: https://github.com/Automattic/jetpack-videopress/compare/v0.32.19...v0.33.0
 [0.32.19]: https://github.com/Automattic/jetpack-videopress/compare/v0.32.18...v0.32.19
 [0.32.18]: https://github.com/Automattic/jetpack-videopress/compare/v0.32.17...v0.32.18
 [0.32.17]: https://github.com/Automattic/jetpack-videopress/compare/v0.32.16...v0.32.17